prevent exponential number of reconnect requests
If connecting fails (the backend isn't running) then both the close and
error events are triggered. This results in two calls to onFail method
which schedules two retry attempts. If they fail both of them schedules
two new retires resulting in four retries. This continues until the browser is
exausted of resources (pretty fast).

This commit tracks if a connection attempt is schedules and prevent new
attempts from being schedules before the current one is finished.