fix resubscribing to a topic
the previous solution forgets to remove the _signature_ from the
registeredTopics object when it removes the listener for that topic.
The next time you try to subscribe the subscribe method just add the
handler to the handler array, but does not set up a new listener.
Removeing the _signature_ from the registered topics solves this.